# backend/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Authentication API",
    "description": "API for user authentication",
    "version": "1.0.0"
  },
  "paths": {
...
    "/auth/google": {
      "get": {
        "tags": ["General"],
        "summary": "Redirect to Google OAuth2 login",
        "description": "Redirects the user to Google's OAuth2 login page to initiate authentication.",
        "operationId": "googleLogin",
        "responses": {
          "302": {
            "description": "Redirects to Google OAuth2 login page."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/auth/google/callback": {
      "get": {
        "tags": ["General"],
        "summary": "Handle OAuth2 callback from Google",
        "description": "Handles the OAuth2 callback from Google and authenticates the user.",
        "operationId": "googleCallback",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "description": "Authorization code received from Google OAuth2.",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "302": {
            "description": "Redirects to the frontend dashboard after successful authentication."
          },
          "400": {
            "description": "Authorization code not provided."
          },
          "500": {
            "description": "Authentication failed."
          }
        }
      }
    },
    "/logout": {
      "get": {
        "tags": ["General"],
        "summary": "Log out the user and clear session",
        "description": "Logs out the user by destroying the session and clearing the JWT cookie.",
        "operationId": "logout",
        "responses": {
          "200": {
            "description": "Successfully logged out."
          },
          "400": {
            "description": "Failed to log out."
          }
        }
      }
    },
...
    "/applications": {
      "get": {
        "tags": ["Applications"],
        "summary": "Fetch all applications",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "List of all applications",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Application"
              }
            }
          },
          "500": {
            "description": "Error fetching applications"
          }
        }
      },
      "post": {
        "tags": ["Applications"],
        "summary": "Create a new application",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{
          "in": "body",
          "name": "application",
          "description": "New application data",
          "schema": {
            "$ref": "#/definitions/ApplicationCreateRequest"
          }
        }],
        "responses": {
          "201": {
            "description": "Application created successfully",
            "schema": {
              "$ref": "#/definitions/Application"
            }
          },
          "500": {
            "description": "Error creating application"
          }
        }
      },
      "options": {
        "tags": ["Applications"],
        "summary": "Fetch allowed methods for applications",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Allowed HTTP methods",
            "schema": {
              "$ref": "#/definitions/AllowedMethodsResponse"
            }
          }
        }
      }
    },
    "/applications/user/{userId}": {
      "get": {
        "tags": ["Applications"],
        "summary": "Fetch applications by user ID",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{
          "name": "userId",
          "in": "path",
          "required": true,
          "type": "integer",
          "description": "User ID"
        }],
        "responses": {
          "200": {
            "description": "List of applications for the user",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Application"
              }
            }
          },
          "500": {
            "description": "Error fetching applications for user"
          }
        }
      }
    },
    "/applications/job/{jobId}": {
      "get": {
        "tags": ["Applications"],
        "summary": "Fetch applications by job ID",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{
          "name": "jobId",
          "in": "path",
          "required": true,
          "type": "integer",
          "description": "Job ID"
        }],
        "responses": {
          "200": {
            "description": "List of applications for the job",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Application"
              }
            }
          },
          "500": {
            "description": "Error fetching applications for job"
          }
        }
      }
    },
    "/applications/status/{status}": {
      "get": {
        "tags": ["Applications"],
        "summary": "Fetch applications by status",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{
          "name": "status",
          "in": "path",
          "required": true,
          "type": "string",
          "enum": ["pending", "accepted", "rejected"],
          "description": "Application status"
        }],
        "responses": {
          "200": {
            "description": "List of applications with given status",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Application"
              }
            }
          },
          "400": {
            "description": "Invalid status value"
          },
          "500": {
            "description": "Error fetching applications by status"
          }
        }
      }
    },
    "/applications/{id}": {
      "put": {
        "tags": ["Applications"],
        "summary": "Update application status",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{
          "name": "id",
          "in": "path",
          "required": true,
          "type": "integer",
          "description": "Application ID"
        }, {
          "in": "body",
          "name": "status",
          "description": "New status for the application",
          "schema": {
            "$ref": "#/definitions/ApplicationStatusUpdateRequest"
          }
        }],
        "responses": {
          "200": {
            "description": "Application status updated successfully"
          },
          "400": {
            "description": "Invalid status value"
          },
          "500": {
            "description": "Error updating application status"
          }
        }
      },
      "delete": {
        "tags": ["Applications"],
        "summary": "Delete an application by ID",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{
          "name": "id",
          "in": "path",
          "required": true,
          "type": "integer",
          "description": "Application ID"
        }],
        "responses": {
          "200": {
            "description": "Application deleted successfully"
          },
          "500": {
            "description": "Error deleting application"
          }
        }
      },
      "patch": {
        "tags": ["Applications"],
        "summary": "Partially update an application",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{
          "name": "id",
          "in": "path",
          "required": true,
          "type": "integer",
          "description": "Application ID"
        }, {
          "in": "body",
          "name": "application",
          "description": "Application fields to update",
          "schema": {
            "$ref": "#/definitions/ApplicationUpdateRequest"
          }
        }],
        "responses": {
          "200": {
            "description": "Application updated successfully"
          },
          "400": {
            "description": "No fields provided for update"
          },
          "404": {
            "description": "Application not found"
          },
          "500": {
            "description": "Error updating application"
          }
        }
      },
      "head": {
        "tags": ["Applications"],
        "summary": "Check if an application exists",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{
          "name": "id",
          "in": "path",
          "required": true,
          "type": "integer",
          "description": "Application ID"
        }],
        "responses": {
          "200": {
            "description": "Application exists"
          },
          "404": {
            "description": "Application not found"
          },
          "500": {
            "description": "Error checking application existence"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Application": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "user_id": {
          "type": "integer"
        },
        "job_id": {
          "type": "integer"
        },
        "application_date": {
          "type": "string",
          "format": "date-time"
        },
        "status": {
          "type": "string",
          "enum": ["pending", "accepted", "rejected"]
        }
      }
    },
    "ApplicationCreateRequest": {
      "type": "object",
      "properties": {
        "user_id": {
          "type": "integer"
        },
        "job_id": {
          "type": "integer"
        }
      },
      "required": ["user_id", "job_id"]
    },
    "ApplicationStatusUpdateRequest": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "enum": ["pending", "accepted", "rejected"]
        }
      },
      "required": ["status"]
    },
    "ApplicationUpdateRequest": {
      "type": "object",
      "properties": {
        "status": {
          "type":"string"
        }
      }
    },
    "AllowedMethodsResponse": {
      "type": "object",
      "properties": {
        "allowed_methods": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "example": ["GET", "POST", "OPTIONS"]
        }
      }
    },
...
  },
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Enter your Bearer token in the format: Bearer token",
      "x-display-name": "Bearer Token",
      "x-token-prefix": "Bearer "
    }
  }
}